+++ tests/tstests/util/common.hpp
@@ -125,7 +125,7 @@ typedef struct TsTestsInfo
 #	else
 	std::function<std::vector<TSTESTS_COORDINATE_TYPE>(uint64_t const)> next_point_getter;
 #	endif // TSTESTS_OPTIMISE_FOR_DIGITAL_NETS
-	std::function<std::vector<std::vector<uint>>(uint const)> gamma_matrix_getter;
+	std::function<std::vector<std::vector<uint>>(uint64_t const)> gamma_matrix_getter;
 	FILE       *log_file;
 }
 TsTestsInfo;
