+++ tests/tstests/util/raref.cpp
@@ -469,12 +469,12 @@ RAREFMatrix cast_matrix(std::vector<std::vector<uint>> const &src)
 	return dst;
 }
 
-TsTestsReturnCode find_defect(uint &ro, uint m, uint s, std::function<std::vector<std::vector<uint>>(uint const)> const &gamma_matrix_getter)
+TsTestsReturnCode find_defect(uint64_t &ro, uint64_t m, uint64_t s, std::function<std::vector<std::vector<uint>>(uint64_t const)> const &gamma_matrix_getter)
 {
 	try
 	{
 		std::vector<RAREFMatrix> genMat;
-		for (uint i = 0; i < s; i++)
+		for (uint64_t i = 0; i < s; i++)
 		{
 			genMat.push_back(cast_matrix(gamma_matrix_getter(i)));
 		}
